[dev] + experimental penalty terms in hybrid opt
Added experimental penalty terms in 2nd stage hybrid optimization. The first one
minimizes peak cycles with a factor to prioritize minimization of base
cycles.
Additionally, a monotonically decreasing factor was introduced and multiplied
with each value in the objective function. This way, inter-storage-power-flow is
made unambiguous as it will reload as late as possible. Also, for ideal
storages, this will have the same effect.

On branch develop
Changes to be committed:
      modified:   batch.py
      modified:   factories.py
      modified:   hybridbuilder.py
      modified:   optimize_ess.py