[dev] * saving unneccessary data to disc
Throwing away all data within python scope and saving them to disc for
archiving. This way, the consumed amount of RAM is hopefully a lot less and
saving and crunching data does not take forever.

On branch develop
Your branch is ahead of 'origin/develop' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
      modified:   .gitignore
      modified:   batch.py
      modified:   hybriddia.py
      modified:   optimize_ess.py
      modified:   powersignal.py
      modified:   results.py