[dev] + burst_split() and equals() for signals
Added a burst split method for signals, which splits a signal into n
signals of equal length.
Added an sig1.equals(sig2) comparision, which compares whole signals,
and not an elementwise comparision as __eq__().
Fixed a bug in time vectors equal length comparision.

On branch develop
Changes to be committed:
      modified:   optess/signal.py