[ideal] + IdealSingle, ~ Target, * Storage
Implemented basic part of IdealSingleOptBuilder. Reworked naming scheme
and expanded functionality. Aim is that the user can choose if he/she wants
to minimize Energy, Power or Target, in a second and third stage, Cycles
and Dynamic Stress can be minimized.
Renamed former Objectives 'Power' (Peak Cutting), 'Energy' (Self
Consumption) and 'Exact' into Target, to make the distinction clear to
minimize Energy, Power or Target -- which is now called Objective.

As this is a only a step within a larger framework change, the code is
likely not working atm.

On branch feature/ideal
Changes to be committed:
      modified:   README.md
      modified:   optess/__init__.py
      modified:   optess/factories.py
      modified:   optess/hybridbuilder.py
      modified:   optess/hybriddia.py
      new file:   optess/idealsingleoptbuilder.py
      deleted:    optess/objective.py
      modified:   optess/optimize_ess.py
      modified:   optess/overdimdia.py
      modified:   optess/singlebuilder.py
      modified:   optess/singledia.py
      modified:   optess/storage.py
      new file:   optess/target.py
      modified:   scripts/batch.py
      modified:   scripts/cyclestudy.py
      modified:   scripts/testenergy.py
      modified:   scripts/testhybconstraint.py
      modified:   scripts/testoverdim.py
      modified:   scripts/testpeh.py
      modified:   scripts/testrefactor.py
      modified:   scripts/testtorque.py