[ideal] + finished single opt, untested
Reworked the design of objective and build functions. Instead of
introducing stages, the building functions now are called min_energy,
min_power and so on. Therefore, the objective class is removed.
Arising question: combine different targets --> peak power cutting as
constraint and self consumption to maximize.

On branch feature/ideal
Changes to be committed:
      modified:   optess/idealsingleoptbuilder.py
      modified:   optess/target.py