[dev] * set the energy for optimisation, untested
Normally, the hybridisation diagram problem would be solved by solving
the single storage problem to get the energy, first. Now, the compute
functions of hybrid dia have an additional optional argument
'singleenergy' to provide this parameter instead of calculating it. This
introduces the possibility to calculate the hybridisation curve for a
subproblem of a larger one with windowing.

With this, the energy content of a storage for a large signal can be
determined the traditional way. Afterwards, the large signal can be
split into multiple smaller signals, which can be solved by themselves
but the needed energy content for the original signal can be passed, so
they are comparable.

On branch develop
Your branch is up to date with 'origin/develop'.

Changes to be committed:
      modified:   optess/hybriddia.py
      modified:   optess/torque.py
      modified:   torque/single.py
      modified:   torque/single.sh