[dev] * torque single.py float conversion
The new argument singleenergy for the single conversion is passed within
the torque system with sys.argv as a string which is then converted to a
float. But the argument is optional, so it can also be 'None' and
float('None') throws an error. So it is checked before if the conversion
is valid, else None is passed.

On branch develop
Your branch is up to date with 'origin/develop'.

Changes to be committed:
      modified:   optess/torque.py
      modified:   torque/single.py