[dev] ~ constraint for base+peak=single, untested
Changed the constraint for the energy capacities within the hybrid
optimization model builder.
Previous constraint was base + peak <= single
new constraint is base + peak == single
This is done to allow investiagions with overdimenisioning in energy
(especially to investigate windowing). The downside is that the hybrid
optimization model is not usable as a standalone class anymore as it is
now mandatory to set the singleenergycapacity in the function call and
to calculate the single storage problem, first.

On branch develop
Your branch is up to date with 'origin/develop'.

Changes to be committed:
      modified:   optess/hybridbuilder.py