[dev] * changed energy in single results, untested
Overwrite the content of hyb.single.dim with the passed arguement for
the set singleenergy value. The _add_extreme_points function will then
correctly load singleenergy value into the inter and nointer curve.
This is done for the window analysis.

Commented out the area points parts within the plotting function, so an
analysis can be performed and plotted without these.

On branch develop
Your branch is up to date with 'origin/develop'.

Changes to be committed:
      modified:   optess/hybriddia.py